refactored login function'
